requirements.txt
duckduckgo-search>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
rich>=13.0.0
click>=8.1.0
pandas>=2.0.0
pyyaml>=6.0.0
jinja2>=3.1.0
tldextract>=5.0.0
urllib3>=2.0.0
flask>=3.0.0
gunicorn>=21.2.0
